extract date functions to date utils
