+++ utils/date-utils.js
@@ -40,3 +40,31 @@ export function formatDate(startDate, endDate, locale) {
 function isValidDate(date) {
   return date instanceof Date && !isNaN(date)
 }
+
+/**
+ *Converts the given date from PDT timezone to the given timezone.
+ *
+ * @param {String} date
+ * @param {String} timezone
+ * @returns {String}
+ */
+export function pdtToLocaleDate(date, timezone) {
+  const datePDT = new Date(`2021-06-08 ${date} PDT`)
+  return datePDT.toLocaleString('en-US', {
+    timeZone: timezone,
+  })
+}
+
+/** Formats a full date to just the time in short format, e.g:
+ * "6/8/2021, 6:00:00 PM" --> "6:00 pm"
+ *
+ * @param {String} date
+ * @returns {String}
+ */
+export function formatTime(time) {
+  const timeSplit = time.split(' ')
+  const hourFull = timeSplit[1] // 6:00:00
+  const hourShort = hourFull.substring(0, hourFull.length - 3) // 6:00
+  const period = timeSplit[2].toLowerCase() // pm
+  return hourShort.concat(' ').concat(period) // 6:00 pm
+}
